+++ src/main/java/mashup/backend/tich/category/service/AdminCategoryService.java
@@ -1,36 +1,20 @@
-package mashup.backend.tich.admin.service;
+package mashup.backend.tich.category.service;
 
 import lombok.RequiredArgsConstructor;
+import mashup.backend.tich.category.domain.Category;
+import mashup.backend.tich.category.domain.CategoryRepository;
+import mashup.backend.tich.category.dto.CategoryResponseDto;
+import mashup.backend.tich.category.dto.CategorySaveRequestDto;
+import mashup.backend.tich.category.dto.CategoryUpdateRequestDto;
 import mashup.backend.tich.exception.ResultDoseNotExistException;
-import mashup.backend.tich.item.domain.Category;
-import mashup.backend.tich.item.domain.CategoryRepository;
-import mashup.backend.tich.item.dto.CategoryResponseDto;
-import mashup.backend.tich.item.dto.CategorySaveRequestDto;
-import mashup.backend.tich.item.dto.CategoryUpdateRequestDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @RequiredArgsConstructor
 @Service
 public class AdminCategoryService {
 
     private final CategoryRepository categoryRepository;
-
-    @Transactional(readOnly = true)
-    public List<CategoryResponseDto> showCategories() {
-        List<Category> categories = categoryRepository.findAll();
-
-        return CategoryResponseDto.listOf(categories);
-    }
-
-    @Transactional(readOnly = true)
-    public CategoryResponseDto showCategory(Long categoryId) {
-        Category category = findCategoryById(categoryId);
-
-        return CategoryResponseDto.of(category);
-    }
 
     @Transactional
     public CategoryResponseDto saveCategory(CategorySaveRequestDto requestDto) {
