+++ src/main/java/mashup/backend/tich/faq/domain/FaqRepository.java
@@ -1,4 +1,4 @@
-package mashup.backend.tich.admin.domain;
+package mashup.backend.tich.faq.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
