+++ src/main/java/mashup/backend/tich/faq/service/AdminFaqService.java
@@ -1,36 +1,20 @@
-package mashup.backend.tich.admin.service;
+package mashup.backend.tich.faq.service;
 
 import lombok.RequiredArgsConstructor;
-import mashup.backend.tich.admin.domain.Faq;
-import mashup.backend.tich.admin.domain.FaqRepository;
-import mashup.backend.tich.admin.dto.FaqResponseDto;
-import mashup.backend.tich.admin.dto.FaqSaveRequestDto;
-import mashup.backend.tich.admin.dto.FaqUpdateRequestDto;
 import mashup.backend.tich.exception.ResultDoseNotExistException;
+import mashup.backend.tich.faq.domain.Faq;
+import mashup.backend.tich.faq.domain.FaqRepository;
+import mashup.backend.tich.faq.dto.FaqResponseDto;
+import mashup.backend.tich.faq.dto.FaqSaveRequestDto;
+import mashup.backend.tich.faq.dto.FaqUpdateRequestDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
 @RequiredArgsConstructor
 @Service
-public class FaqService {
+public class AdminFaqService {
 
     private final FaqRepository faqRepository;
-
-    @Transactional(readOnly = true)
-    public List<FaqResponseDto> showFaqs() {
-        List<Faq> faqs = faqRepository.findAll();
-
-        return FaqResponseDto.listOf(faqs);
-    }
-
-    @Transactional(readOnly = true)
-    public FaqResponseDto showFaq(Long faqId) {
-        Faq faq = findFaqById(faqId);
-
-        return FaqResponseDto.of(faq);
-    }
 
     @Transactional
     public FaqResponseDto saveFaq(FaqSaveRequestDto requestDto) {
