+++ src/main/java/mashup/backend/tich/history/domain/HistoryRepository.java
@@ -1,15 +1,16 @@
-package mashup.backend.tich.item.domain;
+package mashup.backend.tich.history.domain;
 
-        import org.springframework.data.jpa.repository.JpaRepository;
-        import org.springframework.data.jpa.repository.Modifying;
-        import org.springframework.data.jpa.repository.Query;
-        import org.springframework.data.repository.query.Param;
-        import org.springframework.transaction.annotation.Transactional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.transaction.annotation.Transactional;
 
-        import java.util.List;
+import java.util.List;
 
 public interface HistoryRepository extends JpaRepository<History, Long> {
     List<History> findAllByUserIdAndItemId(Long userId, Long itemId);
+
     boolean existsAllByItemId(Long itemId);
 
     @Transactional
